Create Thesis Literature review.docx
For the literature review commit, one will find the background research when it comes to wireless networks. The research started by explaining how networks have become an essential in todays and how wireless networks have numerous benefits comparing to normal wired connections. The Types of networks which were introduced over the years are also identified. In the literature review an explanation of the Raspberry Pi was also given, it was explained how this device is very capable and useful for the projects which I will be working on. After, the uses of drones with networks were presented. One can see how other researchers have implemented the idea of extending networks using drones using different methods. Lastly, the local/foreign rights which apply to using drones in public places. This was important to keep in mind as breaking laws can results into fines.