Create jake_grech research methodology.docx
For the methodology commit, one will find the method that was chosen to be used along with all the devices which will be used to have a working prototype. A more detailed description was given on the Raspberry PI which includes technical information which is crucial to my setup. A step by step of the setup was also given of how it was implemented. Each step which was done was explained to make the reader fully understand why each step was taken. Lastly one will find the testing phase, first the created networks were tested by connecting to them and after concluding that they work I moved on the doing surveys using Ekahau site survey software. One can see an explanation of the software and how surveys were carried out.