Create Results and discussion.docx
For the results and discussion section commit, one will find the relevant results to the study which was just undergone. This consists of 4 key metrics which the network protocols were assessed on and compared with each other to identify which would be most suitable for my study. Before concluding any real-word tests for each metric first theoretical study was done to identify what values I am expecting to see from my setup. After this it the Ekahau site survey software projects which were taken from in the above step are analysed to measure the performance of each protocol. The 4 key metrics which the protocols are analysed are (i) coverage area, (ii) transmission rate, (iii) signal strength and (iv) Efficiency. For each metric graphs are shown of the results and this is compared with the theoretical study to finally conclude if the tests are accurate.