conclusion,intro,abstract
For the Conclusion, introduction, abstract commit one will find the introduction section one will find relevant research on how networks are important and have become an essential utility. One will also find the researched question is also stated in this section. After stating my researched question one will find the motivation for this study. Here it was explained why I have picked this area of study and what got me interested in it. Lastly a context of the research is done by explaining how I am going to be undergoing my research study. In the conclusion section one will find the answer of the research question and if I manage to successfully implement a portable network which can be used on a drone. One can also see the limitations that I have encountered while implementing the system. Lastly one will see the any future recommendations which I suggested to improve the system to be more efficient. The abstract contains a short summary of the whole dissertation. Included in this section are the background research for the study, The aims of what my work will focus on, the methods which were used during the implementation , The results which were obtained during the testing, A conclusion of paper which describer the problems found while implementing such setup and suggested future improvements